requirements.txt
python-telegram-bot
google-generativeai
Pillow
python-dotenv
aiohttp
asyncio